Docs formatting issue resolved
In the docs portion, the entire Authorization module (under the "Extended functions or macros" heading) was not present as a code block. The first (defmodule Authorization do) and last (end) lines were not a part of the code block in the docs page (https://hexdocs.pm/maru/Maru.Helper.html#content).